+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Add `group` to `smooth_lss()` to smooth LSS scores by group.
 * Add `optimize_lss()` as an experimental function.
+* Add `prop_slice` to `textmodel_lss()` an experimental argument.
+* Deprecate `auto_weight` in `textmodel_lss()`.
 
 ## Changes in v1.4.0